GenMS Template/requirements.txt

beautifulsoup4>=4.12.3
docx>=0.2.4
flask>=3.1.0
openai>=1.56.2
python-dotenv>=1.0.0
werkzeug>=3.0.0
gradio-client
